+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This project follows semantic versioning.
 
 ### Unreleased
+- [fixed] Bug when repeating a type in an unnamed variant.
 
 ### 0.1.0 (2023-02-13)
 - [added] Initial `subenum` macro creation.